% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvinceGet.R
\name{buffmeteo}
\alias{buffmeteo}
\title{Buffer intersect with meteorological Stations}
\usage{
buffmeteo(point, Buffer = 5000, projectedepsg = 32632)
}
\arguments{
\item{point}{SpatialPointDataFrame; Indicates the position of the points
which are examined together with the meteorological stations}

\item{Buffer}{numeric; width of the buffer in meters}

\item{projectedepsg}{numeric: EPSG code of the projected Coordinate System. 
In order to compute the distance properly we need to provide a projected coordinate system.
Default is 32632 for UTM 32N}
}
\description{
This function add the possibility to create a buffer around a spatial point
and spatially examine it together with the meteorological stations of the province
South Tyrol. The return can either be the nearest station or a list of the distances
from he point to all station within the buffered area.
}
