% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{CheckGaps}
\alias{CheckGaps}
\title{Check MONALISA NA gaps}
\usage{
CheckGaps(x)
}
\arguments{
\item{x}{a MONALISA Server Response}
}
\description{
Analysis Function for the Server response of a MONALISA Request.
This function simply returns the NA gaps as a table
}
