% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sos4Monalisa_jsonDownload.R
\name{convertDate}
\alias{convertDate}
\title{Convert Data from ... to ...}
\usage{
convertDate(date)
}
\arguments{
\item{date}{date; ... Object}
}
\description{
Convert a Date object formatted in ... Format to a "Ymd" format POSIXct
}
