% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvinceDownload.R
\name{downloadMeteo}
\alias{downloadMeteo}
\title{Download single Meteorological Data}
\usage{
downloadMeteo(dburl = NULL, station_code, sensor_code, datestart,
  dateend, path = "", csv = FALSE)
}
\arguments{
\item{dburl}{URL; URL of the province database. If left empty the standard API will be used}

\item{station_code}{string; Code of the station ("SCODE")}

\item{sensor_code}{string; Abbreviation of the sensor of interest (e.g. "N" for Precipitation)}

\item{datestart}{string; Starting time for the download in "Ymd" Format}

\item{dateend}{string; End time for the download in "Ymd" Format}

\item{path}{string; Specify the output path. If left empty only a object is returned}

\item{csv}{boolean; output as csv?}
}
\description{
Function for downloading the meteorological data provided by the meteorological
service of South Tyrol. NO multiple stations download. For that please use `downloadMeteo2`
}
