% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvinceDownload.R
\name{downloadMeteo2}
\alias{downloadMeteo2}
\title{Download Meteorological Data (v2)}
\usage{
downloadMeteo2(dburl = NULL, station_code, sensor_code, datestart,
  dateend)
}
\arguments{
\item{dburl}{URL; URL of the Province Database. If left empty the original API will be used}

\item{station_code}{string; Station of Interest ("SCODE")}

\item{sensor_code}{string; Abbreviation of the sensor of interest (e.g. "N" for Precipitation)}

\item{datestart}{string; Starting time for the download in "Ymd" Format}

\item{dateend}{string; End time for the download in "Ymd" Format}
}
\description{
Function for accessing the Meteorological Data. This function will replace the
Original Function in the next version. No save to CSV is possible anymore. On the other side
it will be possible to speed up the actual processing times and to make multiple requests at once
}
