% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonalisaDownload.R
\name{downloadMonalisa}
\alias{downloadMonalisa}
\title{Download MONALISA Data}
\usage{
downloadMonalisa(starturl = NULL, datestart, dateend, foi = NULL,
  procedure = NULL, property = NULL, path = "", abbr = F, csv = F)
}
\arguments{
\item{starturl}{URL, Path to the  MONLISA Database. 
If empty the adress of the EURAC Monalisa Database (as available on 30.01.2017) will automatically be pasted.}

\item{datestart}{date, Starting Date required in "Y-m-d H:M" format.}

\item{dateend}{date, End Date required "Y-m-d H:M" format.}

\item{foi}{Character, Input for the FOI. Multiple input possible. Read Notes for further information}

\item{procedure}{Character, Input for the Procedure. Multiple input possible. Read Notes for further information}

\item{property}{Character, Input for the Property. Multiple input possible. Read Notes for further information}

\item{path}{Character, Path for the Output of a RData File with the Respnse from the SOS Server. 
If blank the Output is returned as an object in the R Environment.}

\item{abbr}{Boolean, Do you want the Properties to be automatically abbreviated in the Output? 
If TRUE the Abbreviate() function is applied.}

\item{csv}{Boolean, Additionally Save as csv?}
}
\description{
Core script for handling the data provided in the MONALISA Database
collected and stored by EURAC Research. For more information please visit http://monalisasos.eurac.edu/sos/.
}
\note{
For the Parameters foi, procedures and property the Input can be left empty. This is standard for the function. 
In this case the respective items are listed and can then be selected one by one by digitizing the index number. 

When the three parameters are left empty ("") every element is taken into consideration
}
\examples{

s <- "2016-01-01 00:00"
e <- "2016-12-31 00:00"
mnls_down<-downloadMonalisa(datestart = s,dateend = e)

downloadMonalisa(datestart = s,dateend = e,path= "Tempdir",csv=T)

#END

}
