% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{getDataBase}
\alias{getDataBase}
\title{Overview of the Monalisa Database}
\usage{
getDataBase(url = "")
}
\arguments{
\item{url}{URL; URL of the monalisasos. If empty a default path is used}
}
\description{
Get a table with the overview on the MONALISA Database
}
\details{
This function accesses the Monalisa Database via the API in JSON format
and returns a list of the peatured parameters
}
