% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvinceGet.R
\name{getMeteoInfo}
\alias{getMeteoInfo}
\title{Return Meteo South Tyrol Metainformation}
\usage{
getMeteoInfo(format = "table")
}
\arguments{
\item{format}{string; digit "table" if the output should be a Dataframe or "spatial" for a spatial
output as sf-object}
}
\description{
This function is a further development of both functions 
`getMeteoStat` and `getMeteoSensor`. It unifies both information returning the complete range
of information present in the Open Data Portal South Tyrol.
}
