% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvinceGet.R
\name{getMeteoSensor}
\alias{getMeteoSensor}
\title{List all the properties for one or multiple stations}
\usage{
getMeteoSensor(url = NA, SCODE)
}
\arguments{
\item{url}{URL; URL of the Province Database. If left empty the original API will be used.}

\item{SCODE}{string; Station Code of one Station}
}
\description{
Returns a list af all the sensors and datasets available at one station.
}
