% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvinceGet.R
\name{getMeteoSensor_all}
\alias{getMeteoSensor_all}
\title{List all the available properties at the Province API}
\usage{
getMeteoSensor_all()
}
\description{
Returns a list af all the sensors and datasets available via the API
}
