% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonalisaGet.R
\name{getMonalisaDB}
\alias{getMonalisaDB}
\title{Overview of the Monalisa Database}
\usage{
getMonalisaDB(url = NA, subset = "")
}
\arguments{
\item{url}{URL; URL of the Monalisa SOS API. If empty a default path is used}

\item{subset}{character; The subset of interest. If left empty the whole Data frame will be returned. 
Possible Subsets are "foi","procedure","property"}
}
\description{
This function accesses the Monalisa Database via the API in JSON format
and returns a list of the peatured parameters.
}
\examples{

# Get the Information hosted in the MONALISA Database
mnls<-getMonalisaDB()
mnls2<-getMonalisaDB(subset="property")
mnls3<-getMonalisaDB(subset="station")

# END
}
