% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonalisaGet.R
\name{getMonalisaDB_sub}
\alias{getMonalisaDB_sub}
\title{Overview of the Monalisa Database}
\usage{
getMonalisaDB_sub(subset = "station")
}
\arguments{
\item{subset}{character; the name of the subset that has to be taken.
Digit either "station" for a list of tha available stations or "property" for
a list of the available properties in the database}
}
\description{
This function accesses the Monalisa Database via the API in JSON format
and returns a list of the peatured parameters. In respect to the *getMonalisaDB* function
*getMonalisaDB_sub* returns a subset of the whole database information.
}
