% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonalisaGet.R
\name{getMonalisaStat}
\alias{getMonalisaStat}
\title{Monalisa Stations}
\usage{
getMonalisaStat(db = NULL)
}
\arguments{
\item{db}{MonalisaDB Object; For speeding up processing time a already preset MonalisaDB File
relized with the getMonalisaDB() function can be inserted.}
}
\description{
Returns all the Stations in the MONALISA Database
}
