% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sos4Monalisa_jsonDownload.R
\name{MonalisaDownload}
\alias{MonalisaDownload}
\title{Download MONALISA Data}
\usage{
MonalisaDownload(starturl, datestart, dateend, fois = "", path = "",
  csv = FALSE)
}
\arguments{
\item{starturl}{MONALISA URL path, character}

\item{datestart}{Starting Date required, character (Y-m-d H:M)}

\item{dateend}{End Date required, character (Y-m-d H:M)}

\item{fois}{Aternative Input in case the FOIs for the download are already defined}

\item{path}{Path for the Output. If blank the Output is returne din the R Environment}

\item{csv}{Additional Save as csv?, Boolean}
}
\description{
Downloads the Data from EURACs MONLISA Database
}
