% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sos4Monalisa_jsonDownload.R
\name{MonalisaPlot}
\alias{MonalisaPlot}
\title{Plot the MONALISA Stations}
\usage{
MonalisaPlot(db = "")
}
\arguments{
\item{db}{URL, Path to the  MONLISA Database. If empty the adress will automatically be pasted.}
}
\description{
Function to spatially plot the MONALISA Sations and the respective FOI
to an interactive Leaflet plot. Until now only for visualization purposes
}
