% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonalisaShiny.R
\docType{data}
\name{MonaShiny}
\alias{MonaShiny}
\title{MONALISA Shiny App}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
MonaShiny
}
\description{
This Shiny App accesses the MONSALISA Database via a central API.
With this App it is possible to access, visualize and download the desired data
directly from the server. The facilitated few-click access is designed for the
exploration of all stored and maintained datasets.
}
\keyword{datasets}
