% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvincePlot.R
\name{plotMeteoLeaflet}
\alias{plotMeteoLeaflet}
\title{Plot the Province Stations}
\usage{
plotMeteoLeaflet(
  stations = NULL,
  addPoints = NULL,
  addBuff = F,
  widthBuff = 10000
)
}
\arguments{
\item{stations}{Points; Meteorological Data Points of the province of South Tyrol,
If this is left empty all the Stations will be plotted}

\item{addPoints}{Points; Additional Points on the Leaflet Map. If left empty no Points will be added}

\item{addBuff}{Boolean; add a Buffer to the additional points}

\item{widthBuff}{numeric; The width of a buffer in case it is added}
}
\description{
Function for spatially Plotting the desired Province Stations on a leaflet
template. Besides the plotting of the desired meteorological stations other information
can be added such as additional points, a buffer around certain points
}
