% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonalisaPlot.R
\name{plotMonalisaGG}
\alias{plotMonalisaGG}
\title{Plot MONALISA Request inn GGPlot}
\usage{
plotMonalisaGG(x, stat = "line")
}
\arguments{
\item{x}{tbl, Table with the Value that should be implemented in the}

\item{stat}{chr, You want to plot a "line" (default) or "boxplot"}
}
\description{
Plot the MONALISA Requests either as Line or Boxplot. Both outputs
are displayed as GGPlots
}
