% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonalisaGet.R
\name{setMonalisaURL}
\alias{setMonalisaURL}
\title{Set the MONALISA URL}
\usage{
setMonalisaURL(db = NA)
}
\arguments{
\item{url}{character; Input the URL to the Timeseries of the MONALISA Database.
If this parameter is left empty the standard URL will be set.}
}
\description{
Set the MONALISA URL or leave as default.
}
