% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvinceGet.R
\name{buffmeteo}
\alias{buffmeteo}
\title{Buffer intersect with meteorological Stations}
\usage{
buffmeteo(point, bufferW = 5000, getBufferShp = F, dist = F)
}
\arguments{
\item{point}{Spatial Point; Spatial Point Data Frame indicating the position of the Points
which are examined together with the meteorological stations}

\item{bufferW}{numeric; width of the Buffer in meters}

\item{getBufferShp}{Boolean; return only the buffer(s)?}

\item{dist}{Boolean; shall the distances from the Points to the stations be added to
the output?}
}
\description{
This function add the possibility to create a Buffer around a Spatial point
and spatially examine it together with the meteorological stations of the province
South Tyrol. The return can either be the nearest station or a list of the distances
from he point to all station within the buffered area.
}
