% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sos4Monalisa_jsonDownload.R
\name{convertDate}
\alias{convertDate}
\title{Convert SOS Dates}
\usage{
convertDate(date)
}
\arguments{
\item{date}{date; SOS Date Object}
}
\description{
Convert a Date object formatted in SOS Julian Date Format to a "Ymd" format POSIXct
}
