% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvinceGet.R
\name{getMeteoInfo}
\alias{getMeteoInfo}
\title{Return Meteo Metainformation}
\usage{
getMeteoInfo(url = NA, spatial = F)
}
\arguments{
\item{url}{URL; URL of the Province Database. If left empty the original API will be used.}

\item{spatial}{boolean; Output as Simple Feature (sf) Object?}
}
\description{
This function is a further development of both functions 
`getMeteoStat` and `getMeteoSensor`. It unifies both information returning the complete range
of information present in the Open Data Portal South Tyrol.
}
