% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvinceGet.R
\name{getMeteoStat}
\alias{getMeteoStat}
\title{Retrieve meteorological Stations}
\usage{
getMeteoStat(url = NA, format = "table")
}
\arguments{
\item{url}{URL; URL of the Province Database. If left empty the original API will be used.}

\item{format}{string; digit "table" if the output should be a tibble or "spatial" for a spatial
output as sp-object}
}
\description{
Retreive the meteorological stations within the borders of the autonomous province of South Tyrol, Italy
this script dives the possibility to download the information of these stations as
a table containing all relevant information as well as a spatial in form of an "sp" object.
}
