% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonalisaGet.R
\name{getMonalisaDB}
\alias{getMonalisaDB}
\title{Overview of the Monalisa Database}
\usage{
getMonalisaDB(url = "")
}
\arguments{
\item{url}{URL; URL of the Monalisa SOS API. If empty a default path is used}
}
\description{
This function accesses the Monalisa Database via the API in JSON format
and returns a list of the peatured parameters.
}
