% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonalisaGet.R
\name{getMonalisaDB}
\alias{getMonalisaDB}
\title{Overview of the Monalisa Database}
\usage{
getMonalisaDB(url = NA, subset = "all")
}
\arguments{
\item{url}{URL; URL of the Monalisa SOS API. If left empty a default url for the MONALISA database is used.
The url Paramater is needed until the timeseries path (e.g. ".../sos/api/v1/timeseries/")}

\item{subset}{character; The subset of interest. If left empty the whole Data frame will be returned. 
Possible Subsets are "all","id","foi","procedure","property", "combined" and "geom"}
}
\description{
This function accesses the Monalisa Database via the API in JSON format
and returns a list of the featured parameters.
}
