% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sos4Monalisa_jsonDownload.R
\name{MonalisaDownload}
\alias{MonalisaDownload}
\title{Download MONALISA Data}
\usage{
MonalisaDownload(starturl, datestart, dateend, fois = "", path = "",
  csv = FALSE)
}
\arguments{
\item{starturl}{character, MONALISA URL path}

\item{datestart}{date (Y-m-d H:M), Starting Date required, character}

\item{dateend}{date (Y-m-d H:M), End Date required, character}

\item{fois}{string, Aternative Input in case the FOIs for the download are already defined}

\item{path}{String, Path for the Output. If blank the Output is returned as an object in the R Environment}

\item{csv}{Boolean, Additionally Save as csv?}
}
\description{
URL, Downloads the Data from EURACs MONLISA Database
}
