% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sos4Monalisa_jsonDownload.R
\name{MonalisaDownload}
\alias{MonalisaDownload}
\title{Download MONALISA Data}
\usage{
MonalisaDownload(starturl, datestart, dateend, fois = "", path = "",
  csv = F)
}
\arguments{
\item{starturl}{URL, Path to the  MONLISA Database. If empty the adress will automatically be pasted.}

\item{datestart}{date, Starting Date required in "Y-m-d H:M" format.}

\item{dateend}{date, End Date required "Y-m-d H:M" format.}

\item{fois}{Character, Aternative Input in case the FOIs for the download are already defined.}

\item{path}{Character, Path for the Output. If blank the Output is returned as an object in the R Environment.}

\item{csv}{Boolean, Additionally Save as csv?}
}
\description{
Core script for handling the data provided in the MONALISA Database
collected and stored by EURAC Research. For more information please visit http://monalisasos.eurac.edu/sos/.
}
