% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonalisaPlot.R
\name{MonalisaPlot}
\alias{MonalisaPlot}
\title{Plot the MONALISA Stations}
\usage{
MonalisaPlot(db = NULL)
}
\arguments{
\item{db}{optional DB table; If empty the standard db file fill be generated}
}
\description{
Function to spatially plot the MONALISA Stations and the respective FOI
to a Leaflet plot for better visualization.
}
\examples{
## Standard Plotting wothout Database
MonalisaPlot()
}
